Add documentation

+++ dietr/models/person.py
@@ -5,6 +5,7 @@ class PersonModel:
     with the database.
     '''
     def add_person(self, person):
+        '''Insert a person into the database.'''
         user_id = 1
 
         query = '''INSERT INTO person (account_id, name, url)
@@ -14,9 +15,10 @@ class PersonModel:
         cursor.execute(query, (user_id, person['name'], person['url']))
 
         # Execute query
-        connection.commit()
+        return connection.commit()
 
     def edit_person(self, person):
+        '''Update a person in the database. Person id will be preserved.'''
         user_id = 1
 
         query = '''UPDATE person
@@ -30,23 +32,32 @@ class PersonModel:
                                                               user_id))
 
         # Execute query
-        connection.commit()
+        return connection.commit()
 
     def remove_person(self, id):
+        '''Delete a person from the database. Will also remove related
+        tables.
+        '''
+        user_id = 1
+
         cursor = connection.cursor()
         cursor.execute('''DELETE FROM person
-                                WHERE person.id = %s''', id)
+                                WHERE id         = %s
+                                  AND account_id = %s''', (id, user_id))
 
         cursor.execute('''DELETE FROM person_category_relation
-                                WHERE person_id = %s''', id)
+                                WHERE person_id  = %s
+                                  AND account_id = %s''', (id, user_id))
 
         cursor.execute('''DELETE FROM person_ingredient_relation
-                                WHERE person_ingredient_relation.person_id = %s''', id)
+                                WHERE person_id  = %s
+                                  AND account_id = %s''', (id, user_id))
 
         # Execute query
-        connection.commit()
+        return connection.commit()
 
     def get_person(self, url):
+        '''Fetch a person from the database.'''
         user_id = 1
 
         query = '''SELECT *
@@ -57,11 +68,11 @@ class PersonModel:
         cursor = connection.cursor()
         cursor.execute(query, (user_id, url))
 
-        person = cursor.fetchone()
-
-        return person
+        # Return person
+        return cursor.fetchone()
 
     def get_allergies(self, id):
+        '''Get a list of allergies for a person.'''
         query = '''SELECT category.id, category.name
                      FROM category
                           INNER JOIN person_category_relation
@@ -71,11 +82,11 @@ class PersonModel:
         cursor = connection.cursor()
         cursor.execute(query, id)
 
-        allergens = cursor.fetchall()
-
-        return allergens
+        # Return allergies
+        return cursor.fetchall()
 
     def get_ingredients(self, id):
+        '''Fetch a list of all ingredients from the person.'''
         query = '''SELECT ingredient.id, ingredient.name
                      FROM ingredient
                           INNER JOIN person_ingredient_relation
@@ -85,20 +96,20 @@ class PersonModel:
         cursor = connection.cursor()
         cursor.execute(query, id)
 
-        allergens = cursor.fetchall()
-
-        return allergens
+        # Return ingredients
+        return cursor.fetchall()
 
     def get_persons(self):
+        '''Fetch a list of all persons.'''
         user_id = 1
 
         query = '''SELECT *
                      FROM person
-                    WHERE account_id = %s'''
+                    WHERE account_id = %s
+                 ORDER BY name'''
 
         cursor = connection.cursor()
         cursor.execute(query, user_id)
 
-        persons = cursor.fetchall()
-
-        return persons
+        # Return persons
+        return cursor.fetchall()
